#79: refactoring to better reflect reality and improves navigation around the project.
