+++ message.go
@@ -6,25 +6,6 @@ type Message interface {
 	Type() string
 }
 
-type CollisionMessage struct {
-	Entity *Entity
-	To     *Entity
-}
-
-func (collision CollisionMessage) Type() string {
-	return "CollisionMessage"
-}
-
-type renderChangeMessage struct {
-	entity      *Entity
-	oldPriority PriorityLevel
-	newPriority PriorityLevel
-}
-
-func (renderChangeMessage) Type() string {
-	return "renderChangeMessage"
-}
-
 type MessageManager struct {
 	listeners map[string][]MessageHandler
 }
